+++ spec/grape/dsl/request_response_spec.rb
@@ -93,10 +93,8 @@ module Grape
           expect(subject.content_types).to eq(xml: "application/xml",
                                               serializable_hash: "application/json",
                                               json: "application/json",
-                                              jsonapi: "application/vnd.api+json",
-                                              atom: "application/atom+xml",
-                                              rss: "application/rss+xml",
-                                              txt: "text/plain")
+                                              txt: "text/plain",
+                                              binary: "application/octet-stream")
         end
       end