+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 ============
 
 * Your contribution here.
+* [#745](https://github.com/intridea/grape/pull/745): Removed `atom+xml`, `rss+xml`, and `jsonapi` content-types - [@akabraham](https://github.com/akabraham).
+* [#745](https://github.com/intridea/grape/pull/745): Added `:binary, application/octet-stream` content-type - [@akabraham](https://github.com/akabraham).
 
 0.9.0 (8/27/2014)
 =================